# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ArtisanPath API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: subcategory
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: artisanId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (artisan only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /ratings:
    post:
      operationId: rateProduct
      tags: [ratings]
      summary: Rate a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateProductBody"
      responses:
        "201":
          description: Rating submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
  /ratings/product/{productId}:
    get:
      operationId: getProductRatings
      tags: [ratings]
      summary: Get ratings for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rating"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        artisanId:
          type: integer
        artisanName:
          type: string
        artisanCity:
          type: string
        title:
          type: string
        description:
          type: string
        descriptionHindi:
          type: ["string", "null"]
        descriptionTelugu:
          type: ["string", "null"]
        descriptionTamil:
          type: ["string", "null"]
        category:
          type: string
        subcategory:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        estimatedPriceMin:
          type: number
        estimatedPriceMax:
          type: number
        city:
          type: string
        averageRating:
          type: number
        reviewCount:
          type: integer
        verified:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - artisanId
        - artisanName
        - artisanCity
        - title
        - description
        - category
        - subcategory
        - imageUrl
        - price
        - estimatedPriceMin
        - estimatedPriceMax
        - city
        - averageRating
        - reviewCount
        - verified
        - createdAt
    CreateProductBody:
      type: object
      properties:
        artisanId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        subcategory:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        city:
          type: string
      required:
        - artisanId
        - title
        - description
        - category
        - subcategory
        - imageUrl
        - price
        - city
...
    Rating:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        stars:
          type: integer
        review:
          type: string
        createdAt:
          type: string
      required:
        - id
        - productId
        - userId
        - userName
        - stars
        - review
        - createdAt
    RateProductBody:
      type: object
      properties:
        productId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        stars:
          type: integer
        review:
          type: string
      required:
        - productId
        - userId
        - userName
        - stars
        - review
...